Adds hardcoded offsets for concept demonstration
This commit also introduces some logging that will be pulled out later.
The idea is that given a series of labels along an axis, there ought to
be a way to control which end of the axis the label is rendered at. This
way the labels can be seen in the default rendering.

This commit also adds offsets to correct for some challenges encountered
in getting the xzAxis to pad ticks with an equal(ish) amount of spacing.
Trying to do this using the d3 roundRange() and paddingOuter() features
didn't behave as expected, although that may be a misunderstanding on
my part.