Place tickLabel(s) according to labelPosition
When rendering a tickLabel, take into account the desired placement of
the label - determined by the value of the `labelPosition` variable -
along the axis being rendered.

The default value, `labelPosition = 'proximal'`, will result in the
label being rendered along its domain's origin axis.

Setting the value such that `labelPosition = 'distal'` will result in
the label being drawn offset by the length of its domain's range in the
direction perpendicular to its domain's origin axis.

The internal control variable `labelInset` is used to control the sign
of the calculation involving the `tickPadding` variable, and to control
the inclusion of the `distal` offset in the `translation` transform of
the `tickLabel`.