Add example illustrating distal/proximal labels
This example demonstrates the use of the labelPosition parameter to
control the location where tickLabel(s) are rendered relative to each
tick's domain's origin.

There is a bug in this code in that the labelPosition cannot be updated
for those axes that have already been rendered. Reproduce by unchecking
all filters, changing the state of the 'Distal Labels?' checkbox, and
then rechecking all filters.